Rename sandbox query engine to debug
